feat : studyAsset -> document 용어 통일
