+++ src/app/team/[teamId]/page.tsx
@@ -6,8 +6,8 @@ import { Box, Button, Flex, useBreakpointValue } from '@chakra-ui/react';
 import { useEffect, useState } from 'react';
 import { BsLink45Deg } from 'react-icons/bs';
 
+import { DocumentCardProps } from '@/components/DocumentCard/types';
 import Garden3D from '@/components/Garden3D';
-import { StudyAssetCardProps } from '@/components/StudyAssetCard/types';
 import { StudyCardProps } from '@/components/StudyCard/types';
 import TabButton from '@/components/TabButton';
 import Title from '@/components/Title';
@@ -19,8 +19,8 @@ import NavigationButton from '@/containers/team/NavigationButton';
 import StudyGridView from '@/containers/team/StudyGridView';
 import TeamControlPanel from '@/containers/team/TeamControlPanel';
 import TeamMember from '@/containers/team/teamMember';
+import documentCardData from '@/mocks/documentCard';
 import { gardenInfos1 } from '@/mocks/Garden3D';
-import studyAssetCardData from '@/mocks/studyAssetCard';
 import studyCardData from '@/mocks/studyCard';
 import teamInfoData from '@/mocks/teamInfo';
 
@@ -33,7 +33,7 @@ const Page = ({ params }: { params: { teamId: number } }) => {
 
   const [studyArray, setStudyArray] = useState<StudyCardProps[]>([]);
   const [studyLength, setStudyLength] = useState<number>(0);
-  const [assetArray, setAssetArray] = useState<StudyAssetCardProps[]>([]);
+  const [assetArray, setAssetArray] = useState<DocumentCardProps[]>([]);
   const [assetLength, setAssetLength] = useState<number>(0);
 
   const [isCreateStudyModalOpen, setIsCreateStudyModalOpen] = useState<boolean>(false);
@@ -44,7 +44,7 @@ const Page = ({ params }: { params: { teamId: number } }) => {
       setStudyArray(studyCardData.slice(start, start + CARD_PER_PAGE));
     } else if (category === '학습자료') {
       // TODO: 학습자료 목록 조회하기.
-      setAssetArray(studyAssetCardData.slice(start, start + CARD_PER_PAGE));
+      setAssetArray(documentCardData.slice(start, start + CARD_PER_PAGE));
     }
   };
 
@@ -52,7 +52,7 @@ const Page = ({ params }: { params: { teamId: number } }) => {
     // TODO: 아래의 handleNextClick의 조건문을 기능시키기 위해,
     //       팀 상세 정보 조회 api에서 팀의 스터디와 학습자료 갯수를 받아와야할 것 같습니다.
     setStudyLength(studyCardData.length);
-    setAssetLength(studyAssetCardData.length);
+    setAssetLength(documentCardData.length);
   }, []);
 
   useEffect(() => {
