+++ src/containers/study/DocumentModal/types.ts
@@ -1,11 +1,11 @@
-export interface StudyAssetData {
+export interface DocumentData {
   id: number;
   type: string;
   name: string;
   url: string;
 }
 
-export interface StudyAssetModalProps {
+export interface DocumentModalProps {
   isOpen: boolean;
   setIsModalOpen: React.Dispatch<React.SetStateAction<boolean>>;
   title: string;
