+++ src/containers/team/AssetGridView/types.ts
@@ -1,5 +1,5 @@
-import { StudyAssetCardProps } from '@/components/StudyAssetCard/types';
+import { DocumentCardProps } from '@/components/DocumentCard/types';
 
 export interface AssetGridViewProps {
-  assetArray: StudyAssetCardProps[];
+  assetArray: DocumentCardProps[];
 }
